Move measurement code from Types.hs to Measurement.hs
